+++ src/extend-config/extendHooksConfig.ts
@@ -3,8 +3,10 @@ import { HooksConfig } from "../types/FabloConfigExtended";
 
 const extendHooksConfig = (hooksJson: HooksJson | undefined): HooksConfig => {
   const postGenerate = typeof hooksJson?.postGenerate === "string" ? hooksJson.postGenerate : "";
+  const postStart = typeof hooksJson?.postStart === "string" ? hooksJson.postStart : "";
   return {
     postGenerate,
+    postStart,
   };
 };
 export default extendHooksConfig;