+++ src/setup-docker/templates/fabric-docker.sh
@@ -21,6 +21,9 @@ networkUp() {
   installChannels
   installChaincodes
   notifyOrgsAboutChannels
+  if [ -f "$FABLO_NETWORK_ROOT/hooks/post-start.sh" ]; then
+    ("$FABLO_NETWORK_ROOT/hooks/post-start.sh")
+  fi
   printStartSuccessInfo
 }
 
@@ -46,7 +49,7 @@ elif [ "$1" = "chaincode" ] && [ "$2" = "dev" ]; then
 elif [ "$1" = "chaincode" ] && [ "$2" = "invoke" ]; then
   chaincodeInvoke "$3" "$4" "$5" "$6" "$7"
 elif [ "$1" = "chaincodes" ] && [ "$2" = "list" ]; then
-  chaincodeList "$3" "$4"  
+  chaincodeList "$3" "$4"
 elif [ "$1" = "channel" ]; then
   channelQuery "${@:2}"
 elif [ "$1" = "snapshot" ]; then