+++ src/setup-docker/templates/hooks/post-start.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+# The code from this file is called after Fablo starts the Hyperledger Fabric network
+echo "Executing post-start hook"
+
+<%- hooks.postStart %>