+++ src/types/FabloConfigExtended.ts
@@ -172,6 +172,7 @@ export interface OrdererGroup {
 
 export interface HooksConfig {
   postGenerate: string;
+  postStart: string;
 }
 
 export interface FabloConfigExtended {