+++ src/types/FabloConfigJson.ts
@@ -65,6 +65,7 @@ export interface ChaincodeJson {
 
 export interface HooksJson {
   postGenerate?: string;
+  postStart?: string;
 }
 
 export interface FabloConfigJson {