+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Features
 * Full support for Fabric 3.0.0 and drop solo consensus
   [#513](https://github.com/hyperledger-labs/fablo/pull/513)
+* Add post-start hook functionality
+  [#111](https://github.com/hyperledger-labs/fablo/issues/111)
 
 ## 2.1.0
 
@@ -129,7 +131,7 @@
 ## 0.3.0
 
 ### Features
-* Add [Fablo REST](https://github.com/fablo-io/fablo-rest/) support 
+* Add [Fablo REST](https://github.com/fablo-io/fablo-rest/) support
 * By default all peers are anchor peers
 * Support `postGenerate` hook
 * Added support for [Orderer sharding](https://github.com/hyperledger-labs/fablo/issues/220) (multiple orderer groups).